requirements.txt
altgraph==0.17.4
asn1crypto==1.5.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pefile==2023.2.7
pillow==12.0.0
pycparser==2.23
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyPDF2==3.0.1
pystray==0.19.5
python-dotenv==1.1.1
python-pkcs11==0.9.0
pywin32-ctypes==0.2.3
reportlab==4.4.4
requests==2.32.5
setuptools==80.9.0
six==1.17.0
